Make sure metatensor.{torch}.utils.cmake_prefix_path exists
