+++ python/metatensor-core/tests/utils.py
@@ -1,0 +1,7 @@
+import os
+
+import metatensor
+
+
+def test_cmake_prefix_path():
+    assert os.path.exists(metatensor.utils.cmake_prefix_path)
